Backend/requirements.txt
pymongo==4.11
cloudinary==1.42.1
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0
pinecone-client==5.0.1
scikit-learn==1.6.1
langchain==0.3.14
langchain-community==0.3.14
langchain-groq==0.2.3
sentence-transformers==4.0.2
mlxtend==0.23.1
pandas==2.2.3
pprintpp==0.4.0